pyproject.toml
[tool.poetry]
name = "interview-helper"
version = "0.1.0"
description = "pipline with neuronetworks for interwiew helping"
authors = ["anovikov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
PyAudio = "^0.2.14"
Wave = "^0.0.2"
vosk = "^0.3.45"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
watchdog = "^5.0.3"
websockets = "^13.1"
openai = "^1.50.2"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.5.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
